feat: refactor useLyric to a hook
