+++ src/components/player/PIPLyric.tsx
@@ -5,6 +5,7 @@ import TrackPlayer, { Track } from 'react-native-track-player';
 import { useNoxSetting } from '@stores/useApp';
 import { LyricView } from './Lyric';
 import usePlayerControls from '@components/player/controls/usePlayerControls';
+import useLyric from '@hooks/useRNTPLyric';
 
 const PIPLyricView = () => {
   const [currentTrack, setCurrentTrack] = useState<Track | undefined>(
@@ -15,12 +16,18 @@ const PIPLyricView = () => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const _ = usePlayerControls();
 
+  const usedLyric = useLyric({
+    track: currentTrack,
+    artist: currentTrack?.artist ?? '',
+  });
+
   React.useEffect(() => {
     TrackPlayer.getActiveTrack().then(setCurrentTrack);
   }, [currentPlayingId]);
 
   return currentTrack ? (
     <LyricView
+      usedLyric={usedLyric}
       track={currentTrack}
       artist={'n/a'}
       // HACK: for problems see https://github.com/facebook/react-native/issues/34324
